adds functionality of rendering all cards inside HackContainer
