+++ client/containers/HackContainer.jsx
@@ -1,13 +1,36 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import Hack from '../components/Hack'
 
-const HackContainer = ({ category, setCategory, hacks }) => {
+const HackContainer = ({ category, setCategory, hacks, setHacks }) => {
+
+  async function getHacks() {
+    try {
+      const response = await fetch(`/api/${category}`);
+      // console.log('This is our response before we log data: ', response)
+      const data = await response.json();
+      // console.log('data from getHacks: ', data);
+      setHacks(data);
+    } catch (err) {
+      // console.log('getHacks error! ', err);
+    }
+  }
+
+  // Trigger for page rerender once Category change is detected. //
+  useEffect(() => {
+    getHacks();
+  }, [category]);
 
 
+  
+  const hackItems = [];
+  for (let i = 0; i < hacks.length; i++) {
+    hackItems.push(<Hack hacks={hacks[i]} category={category} setCategory={setCategory}/>);
+  }
 
   return (
     <div>
-      <Hack category={category} setCategory={setCategory} hacks={hacks} />
+      {/* <Hack category={category} setCategory={setCategory} hacks={hacks} /> */}
+      {hackItems}
     </div>
   );
 };
